requirements.txt
web3
hashlib
json
azure-storage-blob
py-solc-x